+++ simple.wat
@@ -1,0 +1,5 @@
+(module
+  (func (export "main")
+        (result i32)
+    i32.const 42
+    return))
