+++ src/lib.rs
@@ -2,5 +2,5 @@ mod app;
 mod code_viewer;
 mod data_provider;
 mod data_provider_twiggy;
-mod functions_exporer;
+mod functions_explorer;
 pub use app::TemplateApp;
